benchmarks/requirements.txt
bmark-py>=0.0.4
pyarrow
pandas
polars
numpy
duckdb